kyc-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
asyncpg==0.28.0
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
boto3==1.28.38
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
fastapi-mail==1.2.8
fastapi-utils==0.2.1
Pillow==10.0.0  # For watermarking